refactor: update TrapFocus name in mui-base/src folder.
