+++ packages/mui-base/src/FocusTrap/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './FocusTrap';
+export * from './FocusTrap';
