+++ packages/mui-base/src/FocusTrap/index.js
@@ -1,0 +1,1 @@
+export { default } from './FocusTrap';
