--- packages/mui-base/src/TrapFocus/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './TrapFocus';
-export * from './TrapFocus';
