--- packages/mui-base/src/TrapFocus/index.js
@@ -1,1 +1,0 @@
-export { default } from './TrapFocus';
